backend/pyproject.toml
[project]
name = "embeddings-backend"
version = "0.1.0"
description = "FastAPI backend for word embeddings visualization"
readme = "README.md"
authors = [
    { name = "Baptiste Gregorutti", email = "baptiste.gregorutti@sorbonne-universite.fr" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "sentence-transformers>=3.0.0",
    "umap-learn>=0.5.5",
    "scikit-learn>=1.0.0",
    "numpy>=1.26.0",
    "pydantic>=2.5.0",
]

[build-system]
requires = ["uv_build>=0.8.20,<0.9.0"]
build-backend = "uv_build"
